Description
This dataset has different TV Airing Brands, their product, and their category. The dataset includes the network through which Ads are airing, types of networks like Cable/ Broadcast, and the show name also on which Ads got aired. The dataset includes the data of Dayparts, Time zone, and the time & date at which Ads got aired. It also includes other data like Pod Position (the lesser the valuable), the duration for which Ads aired on screen, Equivalent sales &, and the total amount spent on the Ads aired.